Use NSIntegerMax as the characterIndexForPoint NSNotFound value
- Cocoa defines NSNotFound as NSIntegerMax; returning NSUIntegerMax is not recognized as the sentinel

+++ src/MewUI/Shared/Platform/MacOS/MacOSWindowInterop.cs
@@ -1618,8 +1618,9 @@ internal static unsafe class MacOSWindowInterop
         // Answering 0 here claims every click sits on character 0, which lets the input method
         // consume/redirect clicks to the wrong place. NSNotFound is the honest "no character" answer
         // until a real screen-point-to-index hit test is wired through the composition editor seam.
+        // Cocoa defines NSNotFound as NSIntegerMax, not NSUIntegerMax.
         MacOSWindowBackend.ImeNativeLogger.Write($"objc characterIndexForPoint view=0x{self:x} pt=({point.x},{point.y}) -> NSNotFound");
-        return ulong.MaxValue;
+        return (ulong)long.MaxValue;
     }
 
     [UnmanagedCallersOnly]
